Change size_t to 64 bit

+++ libc-bottom-half/sources/__wasilibc_real.c
@@ -226,7 +226,7 @@ __wasi_errno_t __wasi_fd_filestat_set_times(
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_fd_pread(int32_t arg0, int64_t arg1, int32_t arg2, int64_t arg3, int64_t arg4) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_fd_pread(int32_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("fd_pread")
 ));
@@ -238,7 +238,7 @@ __wasi_errno_t __wasi_fd_pread(
     __wasi_filesize_t offset,
     __wasi_size_t *retptr0
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_fd_pread((int32_t) fd, (int64_t) iovs, (int32_t) iovs_len, (int64_t) offset, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_fd_pread((int32_t) fd, (int64_t) iovs, (int64_t) iovs_len, (int64_t) offset, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
@@ -255,7 +255,7 @@ __wasi_errno_t __wasi_fd_prestat_get(
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_fd_prestat_dir_name(int32_t arg0, int64_t arg1, int32_t arg2) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_fd_prestat_dir_name(int32_t arg0, int64_t arg1, int64_t arg2) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("fd_prestat_dir_name")
 ));
@@ -265,11 +265,11 @@ __wasi_errno_t __wasi_fd_prestat_dir_name(
     uint8_t * path,
     __wasi_size_t path_len
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_fd_prestat_dir_name((int32_t) fd, (int64_t) path, (int32_t) path_len);
+    int32_t ret = __imported_wasi_snapshot_preview1_fd_prestat_dir_name((int32_t) fd, (int64_t) path, (int64_t) path_len);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_fd_pwrite(int32_t arg0, int64_t arg1, int32_t arg2, int64_t arg3, int64_t arg4) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_fd_pwrite(int32_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("fd_pwrite")
 ));
@@ -281,11 +281,11 @@ __wasi_errno_t __wasi_fd_pwrite(
     __wasi_filesize_t offset,
     __wasi_size_t *retptr0
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_fd_pwrite((int32_t) fd, (int64_t) iovs, (int32_t) iovs_len, (int64_t) offset, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_fd_pwrite((int32_t) fd, (int64_t) iovs, (int64_t) iovs_len, (int64_t) offset, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_fd_read(int32_t arg0, int64_t arg1, int32_t arg2, int64_t arg3) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_fd_read(int32_t arg0, int64_t arg1, int64_t arg2, int64_t arg3) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("fd_read")
 ));
@@ -296,11 +296,11 @@ __wasi_errno_t __wasi_fd_read(
     size_t iovs_len,
     __wasi_size_t *retptr0
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_fd_read((int32_t) fd, (int64_t) iovs, (int32_t) iovs_len, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_fd_read((int32_t) fd, (int64_t) iovs, (int64_t) iovs_len, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_fd_readdir(int32_t arg0, int64_t arg1, int32_t arg2, int64_t arg3, int64_t arg4) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_fd_readdir(int32_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("fd_readdir")
 ));
@@ -312,7 +312,7 @@ __wasi_errno_t __wasi_fd_readdir(
     __wasi_dircookie_t cookie,
     __wasi_size_t *retptr0
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_fd_readdir((int32_t) fd, (int64_t) buf, (int32_t) buf_len, (int64_t) cookie, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_fd_readdir((int32_t) fd, (int64_t) buf, (int64_t) buf_len, (int64_t) cookie, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
@@ -369,7 +369,7 @@ __wasi_errno_t __wasi_fd_tell(
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_fd_write(int32_t arg0, int64_t arg1, int32_t arg2, int64_t arg3) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_fd_write(int32_t arg0, int64_t arg1, int64_t arg2, int64_t arg3) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("fd_write")
 ));
@@ -380,11 +380,11 @@ __wasi_errno_t __wasi_fd_write(
     size_t iovs_len,
     __wasi_size_t *retptr0
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_fd_write((int32_t) fd, (int64_t) iovs, (int32_t) iovs_len, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_fd_write((int32_t) fd, (int64_t) iovs, (int64_t) iovs_len, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_create_directory(int32_t arg0, int64_t arg1, int32_t arg2) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_create_directory(int32_t arg0, int64_t arg1, int64_t arg2) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_create_directory")
 ));
@@ -394,11 +394,11 @@ __wasi_errno_t __wasi_path_create_directory(
     const char *path
 ){
     size_t path_len = strlen(path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_create_directory((int32_t) fd, (int64_t) path, (int32_t) path_len);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_create_directory((int32_t) fd, (int64_t) path, (int64_t) path_len);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_filestat_get(int32_t arg0, int32_t arg1, int64_t arg2, int32_t arg3, int64_t arg4) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_filestat_get(int32_t arg0, int32_t arg1, int64_t arg2, int64_t arg3, int64_t arg4) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_filestat_get")
 ));
@@ -410,11 +410,11 @@ __wasi_errno_t __wasi_path_filestat_get(
     __wasi_filestat_t *retptr0
 ){
     size_t path_len = strlen(path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_filestat_get((int32_t) fd, flags, (int64_t) path, (int32_t) path_len, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_filestat_get((int32_t) fd, flags, (int64_t) path, (int64_t) path_len, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_filestat_set_times(int32_t arg0, int64_t arg1, int32_t arg2, int32_t arg3, int64_t arg4, int64_t arg5, int32_t arg6) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_filestat_set_times(int32_t arg0, int64_t arg1, int32_t arg2, int64_t arg3, int64_t arg4, int64_t arg5, int32_t arg6) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_filestat_set_times")
 ));
@@ -428,11 +428,11 @@ __wasi_errno_t __wasi_path_filestat_set_times(
     __wasi_fstflags_t fst_flags
 ){
     size_t path_len = strlen(path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_filestat_set_times((int32_t) fd, flags, (int64_t) path, (int32_t) path_len, (int64_t) atim, (int64_t) mtim, fst_flags);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_filestat_set_times((int32_t) fd, flags, (int64_t) path, (int64_t) path_len, (int64_t) atim, (int64_t) mtim, fst_flags);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_link(int32_t arg0, int64_t arg1, int32_t arg2, int32_t arg3, int32_t arg4, int64_t arg5, int32_t arg6) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_link(int32_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int32_t arg4, int64_t arg5, int64_t arg6) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_link")
 ));
@@ -446,11 +446,11 @@ __wasi_errno_t __wasi_path_link(
 ){
     size_t old_path_len = strlen(old_path);
     size_t new_path_len = strlen(new_path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_link((int32_t) old_fd, old_flags, (int64_t) old_path, (int32_t) old_path_len, (int32_t) new_fd, (int64_t) new_path, (int32_t) new_path_len);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_link((int32_t) old_fd, old_flags, (int64_t) old_path, (int64_t) old_path_len, (int32_t) new_fd, (int64_t) new_path, (int64_t) new_path_len);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_open(int32_t arg0, int64_t arg1, int32_t arg2, int32_t arg3, int32_t arg4, int64_t arg5, int64_t arg6, int32_t arg7, int64_t arg8) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_open(int32_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int32_t arg4, int64_t arg5, int64_t arg6, int32_t arg7, int64_t arg8) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_open")
 ));
@@ -466,11 +466,11 @@ __wasi_errno_t __wasi_path_open(
     __wasi_fd_t *retptr0
 ){
     size_t path_len = strlen(path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_open((int32_t) fd, dirflags, (int64_t) path, (int32_t) path_len, oflags, fs_rights_base, fs_rights_inheriting, fdflags, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_open((int32_t) fd, dirflags, (int64_t) path, (int64_t) path_len, oflags, fs_rights_base, fs_rights_inheriting, fdflags, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_readlink(int32_t arg0, int64_t arg1, int32_t arg2, int64_t arg3, int32_t arg4, int64_t arg5) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_readlink(int32_t arg0, int64_t arg1, int64_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_readlink")
 ));
@@ -483,11 +483,11 @@ __wasi_errno_t __wasi_path_readlink(
     __wasi_size_t *retptr0
 ){
     size_t path_len = strlen(path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_readlink((int32_t) fd, (int64_t) path, (int32_t) path_len, (int64_t) buf, (int32_t) buf_len, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_readlink((int32_t) fd, (int64_t) path, (int64_t) path_len, (int64_t) buf, (int64_t) buf_len, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_remove_directory(int32_t arg0, int64_t arg1, int32_t arg2) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_remove_directory(int32_t arg0, int64_t arg1, int64_t arg2) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_remove_directory")
 ));
@@ -497,11 +497,11 @@ __wasi_errno_t __wasi_path_remove_directory(
     const char *path
 ){
     size_t path_len = strlen(path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_remove_directory((int32_t) fd, (int64_t) path, (int32_t) path_len);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_remove_directory((int32_t) fd, (int64_t) path, (int64_t) path_len);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_rename(int32_t arg0, int64_t arg1, int32_t arg2, int32_t arg3, int64_t arg4, int32_t arg5) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_rename(int32_t arg0, int64_t arg1, int64_t arg2, int32_t arg3, int64_t arg4, int64_t arg5) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_rename")
 ));
@@ -514,11 +514,11 @@ __wasi_errno_t __wasi_path_rename(
 ){
     size_t old_path_len = strlen(old_path);
     size_t new_path_len = strlen(new_path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_rename((int32_t) fd, (int64_t) old_path, (int32_t) old_path_len, (int32_t) new_fd, (int64_t) new_path, (int32_t) new_path_len);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_rename((int32_t) fd, (int64_t) old_path, (int64_t) old_path_len, (int32_t) new_fd, (int64_t) new_path, (int64_t) new_path_len);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_symlink(int64_t arg0, int32_t arg1, int32_t arg2, int64_t arg3, int32_t arg4) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_symlink(int64_t arg0, int64_t arg1, int32_t arg2, int64_t arg3, int64_t arg4) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_symlink")
 ));
@@ -530,11 +530,11 @@ __wasi_errno_t __wasi_path_symlink(
 ){
     size_t old_path_len = strlen(old_path);
     size_t new_path_len = strlen(new_path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_symlink((int64_t) old_path, (int32_t) old_path_len, (int32_t) fd, (int64_t) new_path, (int32_t) new_path_len);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_symlink((int64_t) old_path, (int64_t) old_path_len, (int32_t) fd, (int64_t) new_path, (int64_t) new_path_len);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_path_unlink_file(int32_t arg0, int64_t arg1, int32_t arg2) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_path_unlink_file(int32_t arg0, int64_t arg1, int64_t arg2) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("path_unlink_file")
 ));
@@ -544,11 +544,11 @@ __wasi_errno_t __wasi_path_unlink_file(
     const char *path
 ){
     size_t path_len = strlen(path);
-    int32_t ret = __imported_wasi_snapshot_preview1_path_unlink_file((int32_t) fd, (int64_t) path, (int32_t) path_len);
+    int32_t ret = __imported_wasi_snapshot_preview1_path_unlink_file((int32_t) fd, (int64_t) path, (int64_t) path_len);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_poll_oneoff(int64_t arg0, int64_t arg1, int32_t arg2, int64_t arg3) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_poll_oneoff(int64_t arg0, int64_t arg1, int64_t arg2, int64_t arg3) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("poll_oneoff")
 ));
@@ -559,7 +559,7 @@ __wasi_errno_t __wasi_poll_oneoff(
     __wasi_size_t nsubscriptions,
     __wasi_size_t *retptr0
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_poll_oneoff((int64_t) in, (int64_t) out, (int32_t) nsubscriptions, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_poll_oneoff((int64_t) in, (int64_t) out, (int64_t) nsubscriptions, (int64_t) retptr0);
     return (uint16_t) ret;
 }
 
@@ -586,7 +586,7 @@ __wasi_errno_t __wasi_sched_yield(
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_random_get(int64_t arg0, int32_t arg1) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_random_get(int64_t arg0, int64_t arg1) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("random_get")
 ));
@@ -595,7 +595,7 @@ __wasi_errno_t __wasi_random_get(
     uint8_t * buf,
     __wasi_size_t buf_len
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_random_get((int64_t) buf, (int32_t) buf_len);
+    int32_t ret = __imported_wasi_snapshot_preview1_random_get((int64_t) buf, (int64_t) buf_len);
     return (uint16_t) ret;
 }
 
@@ -613,7 +613,7 @@ __wasi_errno_t __wasi_sock_accept(
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_sock_recv(int32_t arg0, int64_t arg1, int32_t arg2, int32_t arg3, int64_t arg4, int64_t arg5) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_sock_recv(int32_t arg0, int64_t arg1, int32_t arg2, int64_t arg3, int64_t arg4, int64_t arg5) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("sock_recv")
 ));
@@ -626,11 +626,11 @@ __wasi_errno_t __wasi_sock_recv(
     __wasi_size_t *retptr0,
     __wasi_roflags_t *retptr1
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_sock_recv((int32_t) fd, (int64_t) ri_data, (int32_t) ri_data_len, ri_flags, (int64_t) retptr0, (int64_t) retptr1);
+    int32_t ret = __imported_wasi_snapshot_preview1_sock_recv((int32_t) fd, (int64_t) ri_data, (int64_t) ri_data_len, ri_flags, (int64_t) retptr0, (int64_t) retptr1);
     return (uint16_t) ret;
 }
 
-int32_t __imported_wasi_snapshot_preview1_sock_send(int32_t arg0, int64_t arg1, int32_t arg2, int32_t arg3, int64_t arg4) __attribute__((
+int32_t __imported_wasi_snapshot_preview1_sock_send(int32_t arg0, int64_t arg1, int64_t arg2, int32_t arg3, int64_t arg4) __attribute__((
     __import_module__("wasi_snapshot_preview1"),
     __import_name__("sock_send")
 ));
@@ -642,7 +642,7 @@ __wasi_errno_t __wasi_sock_send(
     __wasi_siflags_t si_flags,
     __wasi_size_t *retptr0
 ){
-    int32_t ret = __imported_wasi_snapshot_preview1_sock_send((int32_t) fd, (int64_t) si_data, (int32_t) si_data_len, (int32_t) si_flags, (int64_t) retptr0);
+    int32_t ret = __imported_wasi_snapshot_preview1_sock_send((int32_t) fd, (int64_t) si_data, (int64_t) si_data_len, (int32_t) si_flags, (int64_t) retptr0);
     return (uint16_t) ret;
 }
